refactor: command 패키지 생성 후 이동
